Scriptable Objects Added up the wazoo
Implemented new SO event system to replace previous.
Basically decomissioned the GameManager, it only is responsible for weapon cooldowns (soon to also be removed) and scene changing actions
Moved a lot of events and actions to be handled by this system, so a lot of code was reworked and removed. Code flows a lot better now I feel