Enemy Waves + Player Skills Updated
Added first iteration of enemy waves, it takes a max enemies that can be alive at that wave, a spawn rate, a list of enemies, and a list of elites, and finally a wave duration. Enemies are spawned at the spawn rate interval, whereas elites are spawned evenly throughout the wave.

Redid player skills to use the same new powerup effect system. Now the powerups can take a generic stat type that both powerupstats and now playerstats extend.